requirements.txt
httpx
matplotlib
cartopy
airportsdata
pillow
